// src/enums/mensages.ts
enum Messages { 
  INTERNAL_SERVER_ERROR = 'Internal Server Error',
  TOKEN_NOT_FOUND = 'Token not found',
  TOKEN_INVALID = 'Invalid token', 
  USER_NOT_FOUND = 'User not found',
  USERNAME_REQUIRED = '"username" is required',
  PASSWORD_REQUIRED = '"password" is required',
  LOGIN_SUCESS = 'Login criado com sucesso',  
  USERNAME_PASSWORD_INVALID = 'Username or password invalid',  
  USERNAME_MUST_BE_STRING = '"username" must be a string',
  USERNAME_VALID_LENGTH = '"username" length must be at least 3 characters long',
  CLASSE_REQUIRED = '"classe" is required',
  CLASSE_MUST_BE_STRING = '"classe" must be a string',
  CLASSE_VALID_LENGTH = '"classe" length must be at least 3 characters long',
  LEVEL_VALID = '"level" must be greater than or equal to 1',
  LEVEL_REQUIRED = '"level" is required',
  LEVEL_MUST_BE_NUMBER = '"level" must be a number',
  PASSWORD_MUST_BE_STRING = '"password" must be a string',
  PASSWORD_VALID_LENGTH = '"password" length must be at least 8 characters long',  
  NAME_REQUIRED = '"name" is required',
  NAME_MUST_BE_STRING = '"name" must be a string',
  NAME_VALID_LENGTH = '"name" length must be at least 3 characters long',
  AMOUNT_REQUIRED = '"amount" is required',
  AMOUNT_MUST_BE_STRING = '"amount" must be a string',
  AMOUNT_VALID_LENGTH = '"amount" length must be at least 3 characters long',
  PRODUCTS_REQUIRED = '"productsIds" is required',
  PRODUCTS_MUST_BE_ARRAY = '"productsIds" must be an array',
  PRODUCTS_VALID_NUMBERS = '"productsIds" must include only numbers',
}
  
export default Messages;